+++ cmd/web/main.go
@@ -1,21 +1,12 @@
 package main
 
 import (
-	"bytes"
-	"embed"
-	"encoding/json"
-	"errors"
-	"io"
 	"log"
 	"net/http"
 	"regexp"
-	"strings"
-	"time"
 
 	"github.com/mbaraa/danklyrics/internal/config"
-	"github.com/mbaraa/danklyrics/pkg/client"
-	"github.com/mbaraa/danklyrics/pkg/provider"
-	website "github.com/mbaraa/danklyrics/website/user"
+	"github.com/mbaraa/danklyrics/internal/handlers/web"
 	"github.com/tdewolff/minify/v2"
 	"github.com/tdewolff/minify/v2/css"
 	"github.com/tdewolff/minify/v2/html"
@@ -26,15 +17,10 @@ import (
 )
 
 var (
-	lyricser *client.Http
-
-	publicFiles embed.FS
-	minifyer    *minify.M
+	minifyer *minify.M
 )
 
 func init() {
-	publicFiles = website.FS()
-
 	minifyer = minify.New()
 	minifyer.AddFunc("text/css", css.Minify)
 	minifyer.AddFunc("text/html", html.Minify)
@@ -42,212 +28,20 @@ func init() {
 	minifyer.AddFuncRegexp(regexp.MustCompile("^(application|text)/(x-)?(java|ecma)script$"), js.Minify)
 	minifyer.AddFuncRegexp(regexp.MustCompile("[/+]json$"), mjson.Minify)
 	minifyer.AddFuncRegexp(regexp.MustCompile("[/+]xml$"), xml.Minify)
-
-	var err error
-	lyricser, err = client.NewHttp(client.Config{
-		Providers:  []provider.Name{provider.Dank, provider.LyricFind},
-		ApiAddress: config.Env().ApiAddress,
-	})
-	if err != nil {
-		panic(err)
-	}
-}
-
-func handleIndex(w http.ResponseWriter, r *http.Request) {
-	if strings.Contains(r.URL.Path, ".go") || strings.Contains(r.URL.Path, "admin") {
-		http.Redirect(w, r, "https://www.youtube.com/watch?v=dQw4w9WgXcQ", http.StatusTemporaryRedirect)
-		return
-	}
-
-	content, err := publicFiles.Open("index.html")
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println(err)
-		return
-	}
-
-	w.Header().Set("Content-Type", "text/html")
-	_, _ = io.Copy(w, content)
-	_ = content.Close()
-}
-
-func handleGetSongLyrics(w http.ResponseWriter, r *http.Request) {
-	artistName, okArtist := r.URL.Query()["artist"]
-	albumName, okAlbum := r.URL.Query()["album"]
-	songName, okSong := r.URL.Query()["song"]
-
-	if !okArtist && !okAlbum && !okSong {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("missing all query parameters `artist`, `album` and `song`"))
-		return
-	}
-
-	if !okSong {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("missing required query parameter `song`"))
-		return
-	}
-
-	searchInput := provider.SearchParams{
-		SongName: songName[0],
-	}
-	if okAlbum {
-		searchInput.AlbumName = albumName[0]
-	}
-	if okArtist {
-		searchInput.ArtistName = artistName[0]
-	}
-
-	lyricsText, err := lyricser.GetSongLyrics(searchInput)
-	if err != nil {
-		log.Println("oppsie doopsie some shit happened", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("No results were found"))
-		return
-	}
-
-	w.Write([]byte(lyricsText.String()))
-}
-
-func makeApiPostRequest[T any](path, token string, body T) error {
-	jsonBody, err := json.Marshal(body)
-	if err != nil {
-		return err
-	}
-
-	req, err := http.NewRequest(http.MethodPost, config.Env().ApiAddress+path, bytes.NewReader(jsonBody))
-	if err != nil {
-		return err
-	}
-
-	req.Header.Set("Authorization", token)
-
-	resp, err := new(http.Client).Do(req)
-	if err != nil {
-		return err
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		return errors.New("api responded with a non 200 status")
-	}
-
-	return nil
-}
-
-func handleAuthSubmitLyrics(w http.ResponseWriter, r *http.Request) {
-	var reqBody struct {
-		Email string `json:"email"`
-	}
-	err := json.NewDecoder(r.Body).Decode(&reqBody)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("invalid body"))
-		return
-	}
-
-	err = makeApiPostRequest("/auth", "", reqBody)
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("Something went wrong"))
-		return
-	}
-}
-
-func handleConfirmAuthSubmitLyrics(w http.ResponseWriter, r *http.Request) {
-	token, tokenExists := r.URL.Query()["token"]
-
-	if !tokenExists {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("missing token"))
-		return
-	}
-
-	err := makeApiPostRequest("/auth/confirm", "", map[string]string{
-		"token": token[0],
-	})
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("Something went wrong"))
-		return
-	}
-
-	http.SetCookie(w, &http.Cookie{
-		Name:    "token",
-		Value:   token[0],
-		Path:    "/",
-		Expires: time.Now().UTC().Add(time.Hour),
-	})
-
-	http.Redirect(w, r, "/", http.StatusTemporaryRedirect)
-}
-
-func handleSubmitLyrics(w http.ResponseWriter, r *http.Request) {
-	sessionToken, err := r.Cookie("token")
-	if err != nil {
-		w.WriteHeader(http.StatusForbidden)
-		w.Write([]byte("No no no, you need to authenticate first!"))
-		return
-	}
-
-	var lyrics struct {
-		SongName   string `json:"song_name"`
-		ArtistName string `json:"artist_name"`
-		AlbumName  string `json:"album_name"`
-		Plain      string `json:"plain_lyrics"`
-	}
-	err = json.NewDecoder(r.Body).Decode(&lyrics)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("invalid request body"))
-		return
-	}
-
-	if lyrics.SongName == "" {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("missing required field `song_name`"))
-		return
-	}
-	if lyrics.ArtistName == "" {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("missing required field `artist_name`"))
-		return
-	}
-	if lyrics.AlbumName == "" {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("missing required field `album_name`"))
-		return
-	}
-	if lyrics.Plain == "" {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("missing required field `plain_lyrics`"))
-		return
-	}
-
-	err = makeApiPostRequest("/dank/lyrics", sessionToken.Value, lyrics)
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("Something went wrong"))
-		return
-	}
 }
 
 func main() {
+	pages := web.NewPages(nil)
 	pagesHandler := http.NewServeMux()
-	pagesHandler.Handle("/static/", http.StripPrefix("/static", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if strings.Contains(r.URL.Path, ".go") || strings.Contains(r.URL.Path, "admin") {
-			http.Redirect(w, r, "https://www.youtube.com/watch?v=dQw4w9WgXcQ", http.StatusTemporaryRedirect)
-			return
-		}
+	pagesHandler.Handle("/static/", pages.StaticFiles())
+	pagesHandler.HandleFunc("/", pages.HandleIndex)
 
-		http.FileServer(http.FS(publicFiles)).ServeHTTP(w, r)
-	})))
-	pagesHandler.HandleFunc("/", handleIndex)
-
+	apis := web.NewApi(nil)
 	apisHandler := http.NewServeMux()
-	apisHandler.HandleFunc("GET /lyrics", handleGetSongLyrics)
-	apisHandler.HandleFunc("POST /lyrics", handleSubmitLyrics)
-	apisHandler.HandleFunc("POST /auth", handleAuthSubmitLyrics)
-	apisHandler.HandleFunc("GET /auth/confirm", handleConfirmAuthSubmitLyrics)
+	apisHandler.HandleFunc("GET /lyrics", apis.HandleGetSongLyrics)
+	apisHandler.HandleFunc("POST /lyrics", apis.HandleSubmitLyrics)
+	apisHandler.HandleFunc("POST /auth", apis.HandleAuthSubmitLyrics)
+	apisHandler.HandleFunc("GET /auth/confirm", apis.HandleConfirmAuthSubmitLyrics)
 
 	applicationHandler := http.NewServeMux()
 	applicationHandler.Handle("/", minifyer.Middleware(pagesHandler))
